requirements.txt
beautifulsoup4==4.10.0
matplotlib==3.4.3
nltk==3.6.3
pandas==1.3.3
requests==2.26.0
scikit-learn==0.24.2
streamlit
seleniumbase
webdriver-manager